Pre Release Update 2
bug fixes to core editor features
tested it on linux
added proper handling to linux in CMakeLists.txt
ui changes for increased visibility
added a picture

+++ src/raynodes/blocks/Pin.h
@@ -158,7 +158,7 @@ struct InputPin final : Pin {
     if constexpr (pt == STRING) {
       return static_cast<const char*>(nullptr);
     } else if constexpr (pt == INTEGER) {
-      return 0LL;
+      return static_cast<int64_t>(0);
     } else if constexpr (pt == BOOLEAN) {
       return false;
     } else if constexpr (pt == FLOAT) {
